Started work on Module 5: property testing
Currently module 5 is only some proof-of-concept code that I threw
together in under an hour, but it works and shows one way that we could
fairly easily implement (at least simple) property-based testing for
sit.
To try it out, just run module_5.py and paste the code printed by it
at the bottom of the test code generated by sit.py. Then compile and run
that test code and you should see "Passed 100 tests." at the end of the
output.